+++ pyrandwalk/pyrandwalk_error.py
@@ -5,16 +5,10 @@
 class pyrandwalkStateError(Exception):
     """State error class."""
 
-    pass
-
 
 class pyrandwalkTransitionsError(Exception):
     """Transitions matrix error class."""
 
-    pass
-
 
 class pyrandwalkVectorError(Exception):
     """Vector error class."""
-
-    pass