+++ CHANGELOG.md
@@ -5,7 +5,21 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-
+## [1.0] - 2021-06-09
+### Added
+- codefactor tests
+- pypi installation guide
+- `Document.ipynb`
+- `Examples.ipynb`
+- `notebook_check.py`
+- `precision` parameter added to `is_prob_dist` function
+- `payoff` attribute
+- `cost` attribute
+- `discount` attribute
+- `best_policy` method
+### Changes
+- `make_prob_dist` bug fixed
+- `final_dist` bug fixed
 ## [0.09] - 2021-05-16
 ### Added
 - `RandomWalk` class
@@ -20,5 +34,6 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - `get_typeof_classes` method
 - `README.md`
 
-[Unreleased]: https://github.com/sadrasabouri/pyrandwalk/compare/v0.09...dev
-
+[Unreleased]: https://github.com/sadrasabouri/pyrandwalk/compare/v1.0...dev
+[1.0]: https://github.com/sadrasabouri/pyrandwalk/compare/v0.09...v1.0
+[0.09]: https://github.com/sadrasabouri/pyrandwalk/compare/9113e4c...v0.09